Misplaced tests
